Make inspector doc links work for inactive systems. Migrate utilities to make more sense and consolidate file locations
