+++ Assets/MixedRealityToolkit/Inspectors/Utilities/MixedRealityStylesUtility.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: f0a3a15cc81d21a4b8ae481fab6d9c04
+guid: 5cccd45d484d4804fa4bb6a9a1bb8830
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {instanceID: 0}
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
