+++ Assets/MixedRealityToolkit/Utilities/MixedRealityOptimizeUtils.cs.meta
@@ -1,11 +1,11 @@
 fileFormatVersion: 2
-guid: 5cccd45d484d4804fa4bb6a9a1bb8830
+guid: c796b99d4988f874f9a54def75101868
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {instanceID: 0}
+  icon: {fileID: 2800000, guid: 961230b29c294bb780054c5d02eb6180, type: 3}
   userData: 
   assetBundleName: 
   assetBundleVariant:
